Revert "SQL docker file update"
This reverts commit ea7d3cb4579865d3d91b5d7ebc68453ba0205e27.

+++ Road Concept Backend/Road Concept API/src/main/java/fr/enssat/lanniontech/api/entities/simulation/Simulation.java
@@ -21,7 +21,7 @@ public class Simulation implements SQLEntity {
     private int mapID; //TODO: Remove it when the FrontEnd is ready to change...
     private MapInfo mapInfo;
     private String creationDate = new SimpleDateFormat("dd-MM-yyyy").format(new Date());
-    private float samplingRate;
+    private int samplingRate;
     private boolean finish;
     private UUID livingFeatureUUID;
     private UUID workingFeatureUUID;
@@ -62,11 +62,11 @@ public class Simulation implements SQLEntity {
         this.name = name;
     }
 
-    public float getSamplingRate() {
+    public int getSamplingRate() {
         return samplingRate;
     }
 
-    public void setSamplingRate(float samplingRate) {
+    public void setSamplingRate(int samplingRate) {
         this.samplingRate = samplingRate;
     }
